moved drawing basic element functions to seperate file
