+++ vpr/src/draw/draw.h
@@ -15,6 +15,7 @@
 #    include "search_bar.h"
 #    include "draw_debug.h"
 #    include "manual_moves.h"
+#	 include "vtr_ndoffsetmatrix.h"
 
 extern ezgl::application::settings settings;
 extern ezgl::application application;
@@ -32,19 +33,15 @@ void alloc_draw_structs(const t_arch* arch);
 void free_draw_structs();
 
 #ifndef NO_GRAPHICS
-void drawroute(enum e_draw_net_type draw_net_type, ezgl::renderer* g);
-
-void draw_get_rr_pin_coords(int inode, float* xcen, float* ycen, const e_side& pin_side);
-void draw_get_rr_pin_coords(const t_rr_node& node, float* xcen, float* ycen, const e_side& pin_side);
-
-
-void draw_x(float x, float y, float size, ezgl::renderer* g);
-
 const ezgl::color SELECTED_COLOR = ezgl::GREEN;
 const ezgl::color DRIVES_IT_COLOR = ezgl::RED;
 const ezgl::color DRIVEN_BY_IT_COLOR = ezgl::LIGHT_MEDIUM_BLUE;
 
 const float WIRE_DRAWING_WIDTH = 0.5;
+
+t_edge_size find_edge(int prev_inode, int inode);
+
+int get_track_num(int inode, const vtr::OffsetMatrix<int>& chanx_track, const vtr::OffsetMatrix<int>& chany_track);
 
 //Returns the drawing coordinates of the specified pin
 ezgl::point2d atom_pin_draw_coord(AtomPinId pin);
@@ -60,6 +57,14 @@ void draw_reset_blk_color(ClusterBlockId blk_id);
 
 ezgl::color to_ezgl_color(vtr::Color<float> color);
 
+bool draw_if_net_highlighted(ClusterNetId inet);
+std::vector<int> trace_routed_connection_rr_nodes(
+    const ClusterNetId net_id,
+    const int driver_pin,
+    const int sink_pin);
+bool trace_routed_connection_rr_nodes_recurr(const t_rt_node* rt_node,
+                                                    int sink_rr_node,
+                                                    std::vector<int>& rr_nodes_on_path);
 void draw_screen();
 
 /* This routine highlights the blocks affected in the latest move      *
