+++ vpr/src/draw/draw_mux.cpp
@@ -23,6 +23,7 @@
 #include "draw.h"
 #include "draw_rr.h"
 #include "draw_xtoy.h"
+#include "draw_basic.h"
 #include "draw_toggle_functions.h"
 #include "draw_triangle.h"
 #include "draw_searchbar.h"
