+++ vpr/src/draw/draw_rr.h
@@ -53,6 +53,6 @@ void draw_expand_non_configurable_rr_nodes_recurr(int from_node,
                                                          std::set<int>& expanded_nodes);
 int draw_check_rr_node_hit(float click_x, float click_y);
 bool highlight_rr_nodes(float x, float y);
-
+void draw_rr_costs(ezgl::renderer* g, const std::vector<float>& rr_costs, bool lowest_cost_first = true);
 #endif /* NO_GRAPHICS */
 #endif /* DRAW_RR_H */
